requirements.txt
google-genai
python-dotenv
requests
openai
protobuf
